+++ src/pywhtml/html_export.py
@@ -65,8 +65,74 @@ strong {font-weight:normal; text-transform: uppercase}
     def get_chapterMapping(self, chId, chapterNumber):
         """Return a mapping dictionary for a chapter section. 
         """
+
+        ROMAN = [
+            (1000, "M"),
+            (900, "CM"),
+            (500, "D"),
+            (400, "CD"),
+            (100, "C"),
+            (90, "XC"),
+            (50, "L"),
+            (40, "XL"),
+            (10, "X"),
+            (9, "IX"),
+            (5, "V"),
+            (4, "IV"),
+            (1, "I"),
+        ]
+
+        def number_to_roman(n):
+
+            result = []
+
+            for (arabic, roman) in ROMAN:
+                (factor, n) = divmod(n, arabic)
+                result.append(roman * factor)
+
+                if n == 0:
+                    break
+
+            return "".join(result)
+
+        TENS = {30: 'thirty', 40: 'forty', 50: 'fifty',
+                60: 'sixty', 70: 'seventy', 80: 'eighty', 90: 'ninety'}
+        ZERO_TO_TWENTY = (
+            'zero', 'one', 'two', 'three', 'four', 'five', 'six', 'seven', 'eight', 'nine', 'ten',
+            'eleven', 'twelve', 'thirteen', 'fourteen', 'fifteen', 'sixteen', 'seventeen', 'eighteen', 'nineteen', 'twenty'
+        )
+
+        def number_to_english(n):
+
+            if any(not x.isdigit() for x in str(n)):
+                return ''
+
+            if n <= 20:
+                return ZERO_TO_TWENTY[n]
+
+            elif n < 100 and n % 10 == 0:
+                return TENS[n]
+
+            elif n < 100:
+                return number_to_english(n - (n % 10)) + ' ' + number_to_english(n % 10)
+
+            elif n < 1000 and n % 100 == 0:
+                return number_to_english(n / 100) + ' hundred'
+
+            elif n < 1000:
+                return number_to_english(n / 100) + ' hundred ' + number_to_english(n % 100)
+
+            elif n < 1000000:
+                return number_to_english(n / 1000) + ' thousand ' + number_to_english(n % 1000)
+
+            return ''
+
         chapterMapping = FileExport.get_chapterMapping(
             self, chId, chapterNumber)
+
+        chapterMapping['ChNumberEnglish'] = number_to_english(
+            chapterNumber).capitalize()
+        chapterMapping['ChNumberRoman'] = number_to_roman(chapterNumber)
 
         if self.chapters[chId].suppressChapterTitle:
             chapterMapping['Title'] = ''